refactor: move motion to core registration module, clean up tests and some imports
